Simplify root post checks when posting
When replying to a thread, validate the root post using `PostStore.GetSingle` instead of `PostStore.Get` as we only need to know that the root post exists, isn't itself a reply, and belongs to the same channel as the newly created post.

The previous call to `PostStore.Get` fetched all the reply posts to that root thread, even though none of this metadata was used.